Feat add cache (#2)
* feat : add cache

* add test
